Use redis/postgres images from quay registry
The current redis and postgresql images aren't using the fully qualified
registry name and use docker.io.
The images used for those app aren't the same than the one used dowstream
and could cause issue with different versions / parameters.

The images from the sclorg namespace on the quay.io registry are the upstream
equivalent to the downstream images for redis and postgresql.

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>