Use initContainers for django migrations
The current django migration workflow is too complicated and splitted
between:
- k8s batch/job (first run)
- k8s exec on API pod

Instead of doing this logic, we can just leverage the initContainers in
the API deployment for running the django migration scripts (and for
the initial data creation as well).

Currently the migration subcommand is executed on all operator runs but
this isn't needed.
The migration subcommand should only has to be executed when the EDA
container image changes so the deployment will be updated (and the
init container triggered again).

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>
(cherry picked from commit 1a6c5909ec001db4c3dfdc0b8b4bfbe6ba39c4e3)