ui: Do not use EDA_SERVER env variable
Not all EDA UI container image are using the environment variable
substituation from the entrypoint script.
This is only working for upstream eda-ui container image which is
based on alpine.

This image executes a script during the container start to:
- subsitute environement variables (like EDA_SERVER)
- move the template file to the final destination.

```
20-envsubst-on-templates.sh: Running envsubst on
/etc/nginx/templates/default.conf.template to /etc/nginx/conf.d/default.conf
```

All of those tasks can be done by the operator so there's no need
to rely on container image mechanism for this since it doesn't work
for another EDA UI container images.

Instead, we can just do the substitution of the EDA_SERVER value as part
of the eda-ui configmap template.
However, we can't bind mount the file as a regular file instead of a template
in /etc/nginx/conf.d instead of /etc/nginx/templates because the EDA UI
container image includes the default configuration file (which includes
the EDA_SERVER environment server).

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>