feat: [AAP-54950] Add IPv6 support
The changes in the deployment allow for gunicorn and daphne to
bind to `::` instead. As IPv6 FE80 Link-Local is always present,
this won't fail. If the pod has a proper IPv6 address in its
interface, this still works, essentially allowing for seamlessly bind
to IPv4 and IPv6 (dual-stack) addresses, while also enabling the
operator to run in single-stack IPv6 scenarios.

We do the same for the configmap that defines the nginx configuration.
Given it already listens to IPv4, we make it listen to IPv6 wildcard
as well, unless explicitly disabled via the `ipv6_disabled` variable.

This also fixes a few typos in some dev/demo EDA CR definitions.